Time Varying Slope
Speed control (Acceleration, Deceleration) using Sparkfun redbot and accelerometer. I have demonstrated it in a closed environment and by taking accelerometer readings (Taw, pitch, and raw) adjusted the speed of the vehicle such that while going on the upward hill it will increase the speed, and while going downwards from the hill the vehicle will decrease the speed.